Fixed PMD warnings and added Javadocs. Refactored XML data type parsing implementation classes.
