+++ metaschema-maven-plugin/src/main/java/gov/nist/secauto/metaschema/maven/plugin/GenerateSchemaMojo.java
@@ -266,6 +266,7 @@ public class GenerateSchemaMojo
     // }
   }
 
+  @SuppressWarnings("PMD.AvoidCatchingGenericException")
   private void performGeneration() throws MojoExecutionException {
     File outputDir = getOutputDirectory();
     if (getLog().isDebugEnabled()) {
