--- schemagen/src/main/java/gov/nist/secauto/metaschema/schemagen/xml/datatype/package-info.java
@@ -1,11 +1,0 @@
-/*
- * SPDX-FileCopyrightText: none
- * SPDX-License-Identifier: CC0-1.0
- */
-
-/**
- * Support for managing Module data type implementations aligned with the XML
- * schema format for use in schema generation.
- */
-
-package gov.nist.secauto.metaschema.schemagen.xml.datatype;
