+++ schemagen/src/main/java/gov/nist/secauto/metaschema/schemagen/xml/impl/XmlMarkupMultilineDatatypeProvider.java
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: CC0-1.0
  */
 
-package gov.nist.secauto.metaschema.schemagen.xml.datatype;
+package gov.nist.secauto.metaschema.schemagen.xml.impl;
 
-public class XmlMarkupLineDatatypeProvider
+public class XmlMarkupMultilineDatatypeProvider
     extends AbstractXmlMarkupDatatypeProvider {
-  private static final String DATATYPE_NAME = "MarkupLineDatatype";
-  private static final String SCHEMA_RESOURCE_PATH = "/schema/xml/metaschema-markup-line.xsd";
+  private static final String DATATYPE_NAME = "MarkupMultilineDatatype";
+  private static final String SCHEMA_RESOURCE_PATH = "/schema/xml/metaschema-markup-multiline.xsd";
 
   @Override
   protected String getSchemaResourcePath() {
