+++ schemagen/src/main/java/gov/nist/secauto/metaschema/schemagen/xml/impl/XmlProseCompositDatatypeProvider.java
@@ -3,11 +3,7 @@
  * SPDX-License-Identifier: CC0-1.0
  */
 
-package gov.nist.secauto.metaschema.schemagen.xml.datatype;
-
-import gov.nist.secauto.metaschema.schemagen.datatype.CompositeDatatypeProvider;
-import gov.nist.secauto.metaschema.schemagen.datatype.IDatatypeContent;
-import gov.nist.secauto.metaschema.schemagen.datatype.IDatatypeProvider;
+package gov.nist.secauto.metaschema.schemagen.xml.impl;
 
 import org.codehaus.stax2.XMLStreamWriter2;
